software enum moved + softwares reordered
